+++ packages/service-library/src/servicelib/fastapi/long_running_tasks/_client.py
@@ -17,7 +17,7 @@ from tenacity.wait import wait_exponential
 from ...long_running_tasks.errors import GenericClientError
 from ...long_running_tasks.models import ClientConfiguration, TaskId, TaskStatus
 
-DEFAULT_HTTP_REQUESTS_TIMEOUT: Final[PositiveFloat] = 15
+_DEFAULT_HTTP_REQUESTS_TIMEOUT: Final[PositiveFloat] = 15
 
 
 logger = logging.getLogger(__name__)
@@ -207,7 +207,7 @@ def setup(
     app: FastAPI,
     *,
     router_prefix: str = "",
-    http_requests_timeout: PositiveFloat = DEFAULT_HTTP_REQUESTS_TIMEOUT,
+    http_requests_timeout: PositiveFloat = _DEFAULT_HTTP_REQUESTS_TIMEOUT,
 ):
     """
     - `router_prefix` by default it is assumed the server mounts the APIs on
