+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/api_client/_public.py
@@ -24,13 +24,15 @@ from servicelib.fastapi.http_client_thin import (
 )
 from servicelib.fastapi.long_running_tasks.client import (
     Client,
-    ProgressCallback,
-    ProgressMessage,
-    ProgressPercent,
     TaskId,
     periodic_task_result,
 )
 from servicelib.logging_utils import log_context, log_decorator
+from servicelib.long_running_tasks.models import (
+    ProgressCallback,
+    ProgressMessage,
+    ProgressPercent,
+)
 from servicelib.utils import logged_gather
 
 from ....core.dynamic_services_settings.scheduler import (
