+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_task.py
@@ -17,8 +17,8 @@ from models_library.service_settings_labels import SimcoreServiceLabels
 from models_library.services_types import ServicePortKey
 from models_library.users import UserID
 from models_library.wallets import WalletID
-from servicelib.fastapi.long_running_tasks.client import ProgressCallback
 from servicelib.fastapi.long_running_tasks.server import TaskProgress
+from servicelib.long_running_tasks.models import ProgressCallback
 
 from ....core.dynamic_services_settings.scheduler import (
     DynamicServicesSchedulerSettings,
