+++ services/director-v2/tests/integration/02/test_dynamic_sidecar_nodeports_integration.py
@@ -52,11 +52,10 @@ from pytest_simcore.helpers.monkeypatch_envs import setenvs_from_dict
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from servicelib.fastapi.long_running_tasks.client import (
     Client,
-    ProgressMessage,
-    ProgressPercent,
     TaskId,
     periodic_task_result,
 )
+from servicelib.long_running_tasks.models import ProgressMessage, ProgressPercent
 from servicelib.progress_bar import ProgressBarData
 from servicelib.sequences_utils import pairwise
 from settings_library.rabbit import RabbitSettings
